Cleaned up refactoring
